Move LookupValues from Wavenet to global torch modules
